changed color scheme and added multithreading to julia

+++ Cursus/Fract-ol/complexe.c
@@ -6,18 +6,18 @@
 /*   By: akabbadj <akabbadj@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/15 16:21:48 by akabbadj          #+#    #+#             */
-/*   Updated: 2023/02/20 04:21:39 by akabbadj         ###   ########.fr       */
+/*   Updated: 2023/02/21 08:25:38 by akabbadj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 
-double add_real(t_complexe *z1, t_complexe *z2)
+float add_real(t_complexe *z1, t_complexe *z2)
 {
     return (z1->real + z2->real);
 }
 
-double add_imag(t_complexe *z1, t_complexe *z2)
+float add_imag(t_complexe *z1, t_complexe *z2)
 {
     return (z1->imag + z2->imag);
 }
@@ -31,18 +31,18 @@ double add_imag(t_complexe *z1, t_complexe *z2)
 //     return(z);
 // }
 
-double multiply_real(t_complexe z1, t_complexe z2)
+float multiply_real(t_complexe z1, t_complexe z2)
 {
     return(z1.real * z2.real - z1.imag * z2.imag);
 }
 
-double multiply_imag(t_complexe z1, t_complexe z2)
+float multiply_imag(t_complexe z1, t_complexe z2)
 {
     return(z1.real * z2.imag + z2.real * z1.imag);
 }
 
 
-double sqrt_root_modulus(t_complexe z)
+float sqrt_root_modulus(t_complexe z)
 {
     return(sqrt(z.real * z.real + z.imag * z.imag)); 
 }
